create a debian container for unifi controller
Build a Debian based container to run the Ubiquiti Unifi Controller from
official repositories.

It is designed to run as a rootless container with `podman`, meaning
it's run as an unprivileged user on the host, but inside the container
it is `root`.

This uses `systemd` inside the container to manage the mongodb database
service as well as the controller itself.

Unifi controller is hard-coded to use mongodb `4`, however this is
no-longer supported on Debian. Later versions (`5` and `6`) do seem to
run fine, so this container includes a hack to remove the mongodb
dependency.